+++ cmd/suggestion/optuna/v1beta1/requirements.txt
@@ -1,4 +1,4 @@
 grpcio>=1.41.1
 protobuf>=4.21.12,<5
 googleapis-common-protos==1.53.0
-optuna==3.3.0
+optuna==3.6.0
