--- cmd/suggestion/skopt/v1beta1/requirements.txt
@@ -1,13 +1,0 @@
-grpcio>=1.41.1
-cloudpickle==0.5.6
-# This is a workaround to avoid the following error.
-# AttributeError: module 'numpy' has no attribute 'int'
-# See more: https://github.com/numpy/numpy/pull/22607
-numpy==1.23.5
-scikit-learn>=0.24.0, <=1.3.0
-scipy>=1.5.4
-forestci==0.3
-protobuf>=4.21.12,<5
-googleapis-common-protos==1.6.0
-scikit-optimize>=0.9.0
-cython>=0.29.24
