+++ manifests/v1beta1/installs/katib-leader-election/katib-config.yaml
@@ -34,7 +34,7 @@ runtime:
     - algorithmName: hyperband
       image: docker.io/kubeflowkatib/suggestion-hyperband:latest
     - algorithmName: bayesianoptimization
-      image: docker.io/kubeflowkatib/suggestion-skopt:latest
+      image: docker.io/kubeflowkatib/suggestion-goptuna:latest
     - algorithmName: cmaes
       image: docker.io/kubeflowkatib/suggestion-goptuna:latest
     - algorithmName: sobol
